fix: add missing variables, update tests
add jwt variables to function calls with auth-decorator to consume the payload
returned from the decorator.

update postman collection so that it tests for proper error code (403) when
credentials are missing instead of wrong error code (401).

update postman collection so that post request of manager provides a list
of ingredients as is expected by the backend.

add test results of the postman collection.